add lang attribute for a11y and rel="noopener noreferrer" for security
technically, the newest versions of chrome (88) will treat
target="_blank" as if it had a rel="noopener" by default, as well as
other browsers having already added this in...

https://chromeunboxed.com/chrome-javascript-block-malicious-redirects-update#:~:text=The%20new%20update%20to%20Chrome,secretly%20being%20taken%20advantage%20of

but if there's somebody with a much older browser, it couldn't hurt to
protect them with this explicitly.

It also makes your lighthouse scores look better, which is a nice plus!